+++ libs/features/pocketbook/src/lib/components/pocketbook-list/pocketbook-list-item/pocketbook-contributors-dialog/pocketbook-contributors-dialog.component.spec.ts
